Implemented Authorization for user.

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,13 +15,13 @@ export class TasksService {
         private taskRepo: TaskRepository
     ) { }
 
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-        return await this.taskRepo.getTasks(filterDto)
+    async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+        return await this.taskRepo.getTasks(filterDto, user)
     }
 
 
-    async getTaskById(id: number): Promise<Task> {
-        const taskFound = await this.taskRepo.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const taskFound = await this.taskRepo.findOne({ where: { id, userId: user.id } });
 
         if (!taskFound) {
             throw new NotFoundException(`Task with ID '${id}' not found`);
@@ -29,20 +30,19 @@ export class TasksService {
     }
 
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepo.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return this.taskRepo.createTask(createTaskDto, user);
     }
 
-    async deleteTaskById(id: number): Promise<void> {
-        const result = await this.taskRepo.delete(id);
-        console.log(result);
+    async deleteTaskById(id: number, user: User): Promise<void> {
+        const result = await this.taskRepo.delete({ id, userId: user.id });
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID '${id}' not found`);
         }
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status
         await task.save()
         return task;
